Add comment on confidence interval calculation
The confidence interval calculation needs to be revised. it should use a t distribution, further references for example in in "What If There Were No Significance Tests?" Edited ByLisa L. Harlow, Stanley A. Mulaik, James H. Steiger; chapter 9; https://doi.org/10.4324/9781315827353 or "A primer on the understanding, use, and calculation of confidence intervals that are based on central and noncentral distributions." Educational and Psychological Measurement 61: 532–574 and Hedges, L. V.  Distribution theory for Glass’s estimator of effect size and related estimators. Journal of Educational Statistics 6: 107–128.